+++ src/com/android/controleestoque/adapter/ListaItem.java
@@ -5,9 +5,9 @@ public class ListaItem {
 	private String tipo;
 	private Integer minimo;
 	private Integer maximo;
-	private Long atual;
+	private Float atual;
 
-	public ListaItem(String nome, Integer min, Integer max, Long valorAtual) {
+	public ListaItem(String nome, Integer min, Integer max, Float valorAtual) {
 		this.nome = nome;
 		this.minimo = min;
 		this.atual = valorAtual;
@@ -23,7 +23,7 @@ public class ListaItem {
 		this.tipo = null;
 	}
 
-	public ListaItem(String nome, String tipo, Long valor) {
+	public ListaItem(String nome, String tipo, Float valor) {
 		this.nome = nome;
 		this.atual = valor;
 		this.tipo = tipo;
@@ -63,11 +63,11 @@ public class ListaItem {
 		this.maximo = maximo;
 	}
 
-	public Long getAtual() {
+	public Float getAtual() {
 		return atual;
 	}
 
-	public void setAtual(Long atual) {
+	public void setAtual(Float atual) {
 		this.atual = atual;
 	}
